Reorganize value-related files
