+++ value/value.ts
@@ -1,5 +1,5 @@
-import { BaseException } from './exceptions';
-import { Formattable, Formatter } from './format';
+import { BaseException } from '../exceptions';
+import { Formattable, Formatter } from '../format';
 
 export class Nil implements Formattable {
   format(_formatter: Formatter): string {
